+++ R/assertions.R
@@ -56,6 +56,22 @@ is_simsx <- function(x, fail = TRUE) {
     } else return(passes)
 }
 
+#' Check if simulations for individual values for x1 are present
+#'   in sim.out
+#' @param x a sim.out method
+#' @param fail logical whether to return an error if simulation range is not
+#'   present.
+
+is_simsx1 <- function(x, fail = TRUE) {
+    passes <- TRUE
+    if (is.null(x$x1)) passes <- FALSE
+    if (isTRUE(fail)) {
+        if (!isTRUE(passes))
+            stop('Simulations for individual fitted values are not present.',
+                call. = FALSE)
+    } else return(passes)
+}
+
 #' Check if simulations for a range of fitted values are present in sim.out
 #' @param x a sim.out method
 #' @param fail logical whether to return an error if simulation range is not
@@ -64,6 +80,21 @@ is_simsx <- function(x, fail = TRUE) {
 is_simsrange <- function(x, fail = TRUE) {
     passes <- TRUE
     if (is.null(x$range)) passes <- FALSE
+    if (isTRUE(fail)) {
+        if (!isTRUE(passes))
+            stop('Simulations for a range of fitted values are not present.',
+                call. = FALSE)
+    } else return(passes)
+}
+
+#' Check if simulations for a range1 of fitted values are present in sim.out
+#' @param x a sim.out method
+#' @param fail logical whether to return an error if simulation range is not
+#'   present.
+
+is_simsrange1 <- function(x, fail = TRUE) {
+    passes <- TRUE
+    if (is.null(x$range1)) passes <- FALSE
     if (isTRUE(fail)) {
         if (!isTRUE(passes))
             stop('Simulations for a range of fitted values are not present.',