+++ R/utils.R
@@ -106,43 +106,56 @@ setval <- function(val, newval) {
   }
 }
 
+
 #' Calculate the reduced dataset to be used in \code{\link{setx}}
 #'
 #' #' This method is used internally
 #'
-#' @param dataset Zelig object data, possibly split to deal with \code{by} argument
+#' @param dataset Zelig object data, possibly split to deal with \code{by}
+#'   argument
 #' @param s list of variables and their tentative \code{setx} values
-#' @param formula a simplified version of the Zelig object formula (typically with 1 on the lhs)
+#' @param formula a simplified version of the Zelig object formula (typically
+#'   with 1 on the lhs)
 #' @param data Zelig object data
 #' @param avg function of data transformations
-#' @return a list of all the model variables either at their central tendancy or their \code{setx} value
-#' @export
+#' @return a list of all the model variables either at their central tendancy or
+#'   their \code{setx} value
+#'
 #' @keywords internal
-#' @author Christine Choirat
+#' @author Christine Choirat and Christopher Gandrud
+#' @export
+
 reduce = function(dataset, s, formula, data, avg = avg) {
-  pred <- try(terms(fit <- lm(formula, data), "predvars"), silent = TRUE)
-  if ("try-error" %in% class(pred)) # exp and weibull
-    pred <- try(terms(fit <- survreg(formula, data), "predvars"), silent = TRUE)
-  dataset <- model.frame(fit)
-  ldata <- lapply(dataset, avg)
-  if (length(s) > 0) {
-    n <- union(as.character(attr(pred, "predvars"))[-1],
-               names(dataset))
-    if (is.list(s[[1]]))
-      s <- s[[1]]
-    m <- match(names(s), n)
-    ma <- m[!is.na(m)]
-    if (!all(complete.cases(m))) {
-      w <- paste("Variable '", names(s[is.na(m)]),
-                 "' not in data set.\n", sep = "")
-      stop(w, call. = FALSE)
+    pred <- try(terms(fit <- lm(formula, data), "predvars"), silent = TRUE)
+    if ("try-error" %in% class(pred)) # exp and weibull
+        pred <- try(terms(fit <- survreg(formula, data), "predvars"), silent = TRUE)
+
+# browser()
+    dataset <- model.frame(fit)
+    test <- dataset
+
+
+
+    ldata <- lapply(dataset, avg)
+    if (length(s) > 0) {
+        n <- union(as.character(attr(pred, "predvars"))[-1], names(dataset))
+        if (is.list(s[[1]])) s <- s[[1]]
+        m <- match(names(s), n)
+        ma <- m[!is.na(m)]
+        if (!all(complete.cases(m))) {
+            w <- paste("Variable '", names(s[is.na(m)]), "' not in data set.\n",
+                     sep = "")
+            stop(w, call. = FALSE)
+        }
+        for (i in seq(n[ma])) {
+            ldata[n[ma]][i][[1]] <- setval(dataset[n[ma]][i][[1]],
+                                           s[n[ma]][i][[1]])
+        }
     }
-    for (i in seq(n[ma]))
-      ldata[n[ma]][i][[1]] <- setval(dataset[n[ma]][i][[1]],
-                                     s[n[ma]][i][[1]])
-  }
-  return(ldata)
+    return(ldata)
 }
+
+
 
 #' Describe Here
 #' @param qi quantity of interest in the discrete case
@@ -162,18 +175,19 @@ statmat <- function(qi) {
 #' Describe Here
 #' @param qi quantity of interest in the discrete case
 #' @param num number of simulations
-#' @return a formatted qi
+#' @return a formatted quantity of interest
 #' @keywords internal
 #' @author Christine Choirat
 statlevel <- function(qi, num) {
     if (is.matrix(qi)){
         #m <- t(apply(qi, 2, table)) / num
         all.levels <- levels(qi)
-        m <- t(apply(qi, 2, function(x) table(factor(x, levels=all.levels)))) / num
-    }else{
+        m <- t(apply(qi, 2, function(x)
+            table(factor(x, levels=all.levels)))) / num
+    } else {
         m <- table(qi) / num
     }
-  return(m)
+    return(m)
 }
 
 #' Pass Quantities of Interest to Appropriate Summary Function
@@ -302,38 +316,45 @@ to_zelig_mi <- function (...) {
 #' @return an \code{mi} object composed of a list of data frames.
 mi <- to_zelig_mi
 
-#' Convert \code{as.factor}s called inside a \code{zelig} call
+#' Conduct variable transformations called inside a \code{zelig} call
 #'
 #' @param formula model formulae
 #' @param data data frame used in \code{formula}
-#' @param check logical whether to just check if a formula contains internally
-#'   called factors and return \code{TRUE} or \code{FALSE}
+#' @param FUN character string of the transformation function
+#' @param check logical whether to just check if a formula contains an
+#'   internally called transformation and return \code{TRUE} or \code{FALSE}
 #' @param f_out logical whether to return the converted formula
 #' @param d_out logical whether to return the converted data frame. Note:
 #'   \code{f_out} must be missing
+#'
 #' @author Christopher Gandrud
 #' @internal
 
-factorize <- function(formula, data, check, f_out, d_out) {
+transformer <- function(formula, data, FUN = 'log', check, f_out, d_out) {
+
+    if (FUN == 'as.factor') FUN_temp <- 'as\\.factor'
+    else FUN_temp <- FUN
+    FUN_str <- sprintf('%s\\(', FUN_temp)
+
     f <- as.character(formula)[3]
     f_split <- unlist(strsplit(f, split = ' '))
-    to_factor <- grep(pattern = 'as.factor', f_split)
+    to_transform <- grep(pattern = FUN_str, f_split)
 
     if (!missing(check)) {
-        if (length(to_factor) > 0) return(TRUE)
+        if (length(to_transform) > 0) return(TRUE)
         else return(FALSE)
     }
 
-    if (length(to_factor) > 0) {
-        to_factor_raw <- f_split[to_factor]
-        to_factor_raw <- gsub('as\\.factor\\(', '', to_factor_raw)
-        to_factor_plain <- gsub('\\)', '', to_factor_raw)
+    if (length(to_transform) > 0) {
+        to_transform_raw <- f_split[to_transform]
+        to_transform_plain <- gsub(FUN_str, '', to_transform_raw)
+        to_transform_plain <- gsub('\\)', '', to_transform_plain)
 
-        if (!all(to_factor_plain %in% names(data)))
-            stop('Unable to find variable to convert to factor.')
+        if (!all(to_transform_plain %in% names(data)))
+            stop('Unable to find variable to transform.')
 
         if (!missing(f_out)) {
-            f_split[to_factor] <- to_factor_plain
+            f_split[to_transform] <- to_transform_plain
             f_comb <- paste(f_split, collapse = ' ')
             dv <- gsub('\\(\\)', '', formula[2])
             f_new <- paste(dv, '~', f_comb, collapse = ' ')
@@ -341,15 +362,19 @@ factorize <- function(formula, data, check, f_out, d_out) {
             return(f_out)
         }
         else if (d_out) {
-            for (i in to_factor_plain) data[, i] <- as.factor(data[, i])
+            for (i in to_transform_plain)
+              data[, i] <- eval(parse(text = sprintf('%s(data[, i])',
+                                                     FUN)))
             return(data)
         }
     }
-    else if (length(to_factor) == 0) {
-          if (!missing(f_out)) return(formula)
-          else if (d_out) return(data)
+    else if (length(to_transform) == 0) {
+        if (!missing(f_out)) return(formula)
+        else if (d_out) return(data)
     }
 }
+
+
 
 #' Remove package names from fitted model object calls.
 #'
@@ -382,3 +407,17 @@ se_pull <- function(x) {
   return(p_values)
 }
 
+#' Drop intercept columns from a data frame of fitted values
+#'
+#' @param x a data frame
+#' @internal
+
+rm_intercept <- function(x) {
+    intercept_names <- c('(Intercept)', 'X.Intercept.')
+    names_x <- names(x)
+    if (any(intercept_names %in% names(x))) {
+        keep <- !(names(x) %in% intercept_names)
+        x <- x[, names_x[keep]]
+    }
+    return(x)
+}